requirements.txt
torch
opencv_python
PyYAML
h5py
tensorboardX
scipy
sklearn